create seeders

+++ database/migrations/2017_08_23_152023_create_buildings_table.php
@@ -18,7 +18,7 @@ class CreateBuildingsTable extends Migration
 
             $table->string('name');
             $table->string('address');
-            $table->integer('value');
+            $table->bigInteger('value');
             $table->string('function');
 
             //location
